Add '/' to Search URLs
`base_url` is a dot (.) and doc.location does not have a forward slash (/) at the beginning, so this causes the search URLs to be invalid. This fix modifies the windmill theme further but solves the search failure issue.

Another possible solution is updating the base_url, but that might have unintended consequences elsewhere.